requirements.txt
requests==2.31.0
PyYAML==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1
click==8.1.7
colorama==0.4.6
jsonschema==4.20.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
retrying==1.3.4
tenacity==8.2.3
loguru==0.7.2
pydantic==2.5.0
aiohttp==3.9.1
asyncio-throttle==1.0.2 